refactor: rename CardEmpresa to CompanyCard
